requirements.txt
tensorflow==2.19.0
scikit-learn
numpy
pandas
seaborn
streamlit
ipykernel
matplotlib
opencv-python
altair
openai
Werkzeug
Flask
python-dotenv
markdown
gunicorn
langchain_groq